refactor: 도서 관련 타입 변경 사항 적용
- image -> cover로 프로퍼티 이름 변경
- 도서 관련 요청 응답 타입 반영
